requirements.txt
elasticsearch
urllib3
tqdm
python-dotenv